==> 1_scraping/2_scripts_python/4_socioeconomic.ipynb <==
# Collecting socioeconomic and electoral data

### Prerequisites

# Libraries

import pandas as pd # data wrangling
import numpy as np # math operations
import math # math operations
import os # directories
import time # system time
import random # random number generation
import pickle # data compression
import re # regular expressions
import unidecode # regular expressions

import urllib.request # scraping
import requests # scraping
from bs4 import BeautifulSoup # scraping
import ctypes # interface to C
import tweepy # twitter 

import sys # system limit (preventing infinite running)
sys.setrecursionlimit(100000)

import selenium # chrome driver
from selenium import webdriver # chrome driver
import selenium.common.exceptions as selexcept # exception handling

### Socioeconomic data

# Retrieve socioeconomic data from URL, save as df and discard first 6 empty rows

url_strukturdaten = "https://www.bundeswahlleiter.de/dam/jcr/f7566722-a528-4b18-bea3-ea419371e300/btw17_strukturdaten.csv"

socioeconomic_df = pd.read_csv(
    url_strukturdaten, encoding="ISO-8859-1", 
    delimiter=';', 
    decimal=",")
socioeconomic_df.columns = socioeconomic_df.iloc[7]
socioeconomic_df = socioeconomic_df.iloc[8:] 
socioeconomic_df = socioeconomic_df.reset_index(drop=True) 

# Drop redundant columns, NaN rows and rows with state-wide total votes

socioeconomic_df = socioeconomic_df.drop(socioeconomic_df.columns[[7, 8, 51]], axis=1)
socioeconomic_df['Wahlkreis-Nr.'] = socioeconomic_df['Wahlkreis-Nr.'].apply(lambda x: int(x)) 
mask1 = socioeconomic_df[socioeconomic_df['Wahlkreis-Nr.'] > 900].index
mask2 = socioeconomic_df[socioeconomic_df['Wahlkreis-Nr.'].isnull()].index
socioeconomic_df = socioeconomic_df.drop(mask1 | mask2)

socioeconomic_df.head()

new_colnames = [
    'bundesland', 
    'wahlkreis_nr', 
    'wahlkreis', 
    'n_municipalities', 
    'area_sqkm',
    'pop_k', 
    'pop_german_k', 
    'net_births_per_k', 
    'net_migration_per_k', 
    'share_u18', 
    'share_0_18', 
    'share_25_34', 
    'share_35_59', 
    'share_60_74', 
    'share_75_inf', 
    'share_pop_nomigration',
    'share_pop_migration', 
    'share_catholic', 
    'share_protestant',
    'share_other_religion', 
    'share_homeowners', 
    'newly_built_appartments_per_k', 
    'appartments_per_k', 
    'household_income_per_capita', 
    'bip_per_capita',
    'cars_per_k',
    'graduates_berufsschule_per_k',
    'graduates_allgemeinbildend_per_k',
    'share_no_graduation',
    'share_hauptschulabschluss',
    'share_mittlere_reife',
    'share_fachhochschulreife', 
    'children_childcare_per_k',
    'businesses_per_k',
    'craft_businesses_per_k',
    'workforce_per_k',
    'share_workforce_primary_sector',
    'share_workforce_industrial_sector',
    'share_workforce_trade_hospiality',
    'share_workforce_service_public_private',
    'share_workforce_service_misc',
    'welfare_recipients_per_k',
    'share_welfare_recipients_unemployable',
    'share_welfare_recipients_foreign',
    'unemployment_rate',
    'share_unemployment_rate_female',
    'share_unemployment_rate_male',
    'share_unemployment_rate_15_19',
    'share_unemployment_rate_55_64'
]

# Rename columns

socioeconomic_df.columns = new_colnames

# Convert columns w/ numbers to numeric

non_numeric_columns = [0, 1, 2]
numeric_columns = [
    element for element in range(len(socioeconomic_df.columns)) if element not in non_numeric_columns]
for i in numeric_columns:
    socioeconomic_df[socioeconomic_df.columns[i]] = \
    socioeconomic_df[socioeconomic_df.columns[i]].apply(
        lambda x: x.replace(',','.')).astype('float', errors = 'ignore')

# note: " - " is represented as square for column Wahlkreis-Name, coded as " \x96 "; 
# need to address if column is to be used

### Election results (2017 federal elections)

url_zweitstimmen = "https://www.bundeswahlleiter.de/dam/jcr/72f186bb-aa56-47d3-b24c-6a46f5de22d0/btw17_kerg.csv"

zweitstimmen_df = pd.read_csv(
    url_zweitstimmen, 
    encoding="ISO-8859-1", 
    delimiter=';', 
    skiprows=5)

# Drop redundant rows and columns

zweitstimmen_df = zweitstimmen_df.iloc[2:]
zweitstimmen_df = zweitstimmen_df.reset_index(drop=True)
relevant_columns = [0, 1, 2, 17, 21, 25, 29, 33, 37, 41, 45] 
zweitstimmen_df = zweitstimmen_df.iloc[:, relevant_columns]

# Rename columns

new_colnames = [
    'wahlkreis_nr', 
    'district', 
    'bundesland_nr', 
    'zweitstimmen_total', 
    'cdu', 
    'spd', 
    'linke', 
    'gruene', 
    'csu', 
    'fdp', 
    'afd'
]

zweitstimmen_df.columns = new_colnames

numeric_columns = [
    'wahlkreis_nr', 
    'bundesland_nr', 
    'zweitstimmen_total', 
    'cdu', 
    'spd', 
    'linke', 
    'gruene', 
    'csu', 
    'fdp', 
    'afd']

for i in numeric_columns:
    zweitstimmen_df[i] = pd.to_numeric(zweitstimmen_df[i], errors = 'coerce')

# Filter out NaN-rows and rows with state-wide total votes

mask1 = zweitstimmen_df[zweitstimmen_df['bundesland_nr'] == 99.0].index
mask2 = zweitstimmen_df[zweitstimmen_df['bundesland_nr'].isnull()].index
zweitstimmen_df = zweitstimmen_df.drop(mask1 | mask2)

# Merge CDU and CSU votes

zweitstimmen_df['cdu_csu'] = np.where(
    zweitstimmen_df['csu'].isnull(), 
    zweitstimmen_df['cdu'], 
    zweitstimmen_df['csu'])
zweitstimmen_df = zweitstimmen_df.drop(columns = ['cdu', 'csu'])

# Convert votes to shares

zweitstimmen_df['cdu_csu'] = zweitstimmen_df['cdu_csu'] / zweitstimmen_df['zweitstimmen_total']
zweitstimmen_df['spd'] = zweitstimmen_df['spd'] / zweitstimmen_df['zweitstimmen_total']
zweitstimmen_df['linke'] = zweitstimmen_df['linke'] / zweitstimmen_df['zweitstimmen_total']
zweitstimmen_df['gruene'] = zweitstimmen_df['gruene'] / zweitstimmen_df['zweitstimmen_total']
zweitstimmen_df['fdp'] = zweitstimmen_df['fdp'] / zweitstimmen_df['zweitstimmen_total']
zweitstimmen_df['afd'] = zweitstimmen_df['afd'] / zweitstimmen_df['zweitstimmen_total']

zweitstimmen_df.head()

### Merging socioeconomic and election data

socioeconomics_zweitstimmen_df = socioeconomic_df.join(
    zweitstimmen_df.set_index('wahlkreis_nr'), 
    on='wahlkreis_nr')

socioeconomics_zweitstimmen_df.head()

socioeconomics_zweitstimmen_df.columns

# Save output

with open('../3_output/socioeconomics_zweitstimmen.pickle', 'wb') as handle:
    pickle.dump(socioeconomics_zweitstimmen_df, handle, protocol=pickle.HIGHEST_PROTOCOL)

socioeconomics_zweitstimmen_df.to_csv(
    '../3_output/socioeconomics_zweitstimmen.csv', index=False, encoding='utf-8-sig')
